Several improvements to our logging facilities.
Added an `info()` log level that includes a timestamp.

Switched from cli-color to chalk, for no other reason than
that Sindre is awesome.

Added text labels to all the log levels so we signal error,
warning, and success with something in addition to color for
all the color-deficient vision peeps out there.