Command-line arguments can now be arrays or objects.
If a string argument is enclosed in {} or [], it is parsed
as JSON. If the parse is unsuccessful, the original string is
passed as the argument. If the parse succeeds, the created object
is passed along as an array.

This allows us to send either a list of validate hosts (for instance)
or a single one as a string. It is the responsibility of the consumer
of an option type like this to know what to do with the argument in
either format.